requirements.txt
python-twitter
PyYAML
